Use cmov for cselect and cswap
This replaces use of the xor swap with cmov, which is considerably
faster on my Skylake machine -- around ~2800 cycles faster for shared
secret calculation.

The shared secret function does not need a full swap, but rather only a
selection, so we build a function that does only that. The key
generation function does need a full swap, however. In both cases, cmov
is faster.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>