Use correct calculation for P-S
The code used P+S instead of P-S. That's probably fine, having the same
order, but it doesn't match the description in the paper or proof
scripts. Another way of viewing the same thing would be to say that the
code used the negative Y coordinate instead of the positive one
mentioned in the paper.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>